Migrating leftover samples
